Refactor code to new location
