+++ core/errors.py
@@ -1,0 +1,4 @@
+class ValidationError(Exception):
+    def __init__(self, message):
+        super().__init__(message)
+        self.message = message
